Adding extra effects
Unfortunetly the fancy blur effect works as shit
outside Chrome in Desktops. So, I implemented an
opacity approach, which is not so fancy but saves
CPU. Also, an additional "always blur" for mobile
devices. The idea is to have multiple built sites and
configure Nginx or some Node.js to parse the
correct build.